+++ src/test/java/io/github/timoa/misc/JacksonImportStyleTest.java
@@ -1,18 +1,14 @@
-package io.github.timoa;
+package io.github.timoa.misc;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
+import org.openrewrite.Issue;
 import org.openrewrite.java.JavaParser;
 import org.openrewrite.java.OrderImports;
-import org.openrewrite.java.style.ImportLayoutStyle;
-import org.openrewrite.style.NamedStyles;
 import org.openrewrite.test.RecipeSpec;
 import org.openrewrite.test.RewriteTest;
 
-import java.util.Arrays;
-
-import static java.util.Collections.*;
-import static org.openrewrite.Tree.randomId;
+import static java.util.Collections.singletonList;
 import static org.openrewrite.java.Assertions.java;
 
 class JacksonImportStyleTest implements RewriteTest {
@@ -24,64 +20,7 @@ class JacksonImportStyleTest implements RewriteTest {
 
     @DocumentExample
     @Test
-    //this test does not use my defined style but creates one on the fly
-    void inlineStyle() {
-        rewriteRun(
-          spec -> spec.parser(JavaParser.fromJavaVersion().styles(
-            singletonList(
-              new NamedStyles(
-                randomId(),
-                "custom",
-                "custom style",
-                null,
-                emptySet(),
-                singletonList(
-                  ImportLayoutStyle.builder()
-                    .importAllOthers()
-                    .blankLine()
-                    .importPackage("com.fasterxml.jackson.core.*")
-                    .importPackage("tools.jackson.*")
-                    .blankLine()
-                    .importStaticAllOthers()
-                    .build()
-                )
-              )
-            ))),
-          java(
-            """
-              package com.fasterxml.jackson.core.write;
-              
-              import com.fasterxml.jackson.core.*;
-              
-              import java.io.*;
-              
-              import static org.junit.jupiter.api.Assertions.*;
-              
-              import org.junit.jupiter.api.Test;
-              
-              class Test {
-              }
-              """,
-            """
-              package com.fasterxml.jackson.core.write;
-              
-              import java.io.*;
-              import org.junit.jupiter.api.Test;
-              
-              import com.fasterxml.jackson.core.*;
-              
-              import static org.junit.jupiter.api.Assertions.*;
-              
-              class Test {
-              }
-              """
-          )
-        );
-    }
-
-    @Test
-    //This tests seeks to use the style I defined. not sure if it is done correctly
-    void myStyle() {
+    void simpleExample() {
         rewriteRun(
           spec -> spec.parser(JavaParser.fromJavaVersion().styles(
             singletonList(
@@ -158,6 +97,38 @@ class JacksonImportStyleTest implements RewriteTest {
               import static org.junit.jupiter.api.Assertions.*;
               
               class GeneratorBasicDerivedTest {
+              }
+              """
+          )
+        );
+    }
+
+    @Issue("https://github.com/openrewrite/rewrite/issues/4165")
+    //@Test
+    // todo reactivate once the issue above is resolved
+    void shouldNotIndentExtends() {
+        rewriteRun(
+          spec -> spec.parser(JavaParser.fromJavaVersion().styles(
+            singletonList(
+              new JacksonImportStyle())
+          )),
+          java("class BaseClass {}"),
+          java(
+            """              
+              import static org.junit.jupiter.api.Assertions.*;
+
+              import java.io.*;
+              class Test
+                  extends BaseClass {
+              }
+              """,
+            """
+              import java.io.*;
+              
+              import static org.junit.jupiter.api.Assertions.*;
+              
+              class Test
+                  extends BaseClass {
               }
               """
           )
